+++ compiler/src/org.graalvm.compiler.truffle.compiler/src/org/graalvm/compiler/truffle/compiler/phases/inlining/GraphManager.java
@@ -47,13 +47,14 @@ import jdk.vm.ci.meta.ResolvedJavaMethod;
 final class GraphManager {
 
     private final PartialEvaluator partialEvaluator;
-    private final EconomicMap<ResolvedJavaMethod, EncodedGraph> graphCacheForInlining = EconomicMap.create();
+    private final EconomicMap<ResolvedJavaMethod, EncodedGraph> graphCacheForInlining;
     private final EconomicMap<CompilableTruffleAST, GraphManager.Entry> irCache = EconomicMap.create();
     private final PartialEvaluator.Request rootRequest;
 
     GraphManager(PartialEvaluator partialEvaluator, PartialEvaluator.Request rootRequest) {
         this.partialEvaluator = partialEvaluator;
         this.rootRequest = rootRequest;
+        this.graphCacheForInlining = partialEvaluator.getOrCreateEncodedGraphCache();
     }
 
     Entry pe(CompilableTruffleAST truffleAST) {